requirements.txt
schedule==1.2.2
selenium==4.35.0
python-dotenv==1.1.1